Length improvements
- moved all public parse in and from methods Length sub classes to Length
