+++ src/main/java/walkingkooka/tree/text/NoneLength.java
@@ -27,24 +27,13 @@ import java.util.Objects;
 import java.util.Optional;
 
 /**
- * A normal measurement.
+ * A none measurement.
  */
 public final class NoneLength extends Length<Void> implements HasJsonNode, Value<Void> {
 
     final static String TEXT = "none";
 
-    /**
-     * Parses text that contains a normal literal.
-     */
-    public static NoneLength parseNormal(final String text) {
-        CharSequences.failIfNullOrEmpty(text, "text");
-
-        return parseNormal0(text);
-    }
-
-    static NoneLength parseNormal0(final String text) {
-        checkText(text);
-
+    static NoneLength parseNone0(final String text) {
         if(!TEXT.equals(text)) {
             throw new IllegalArgumentException("Invalid normal text " + CharSequences.quoteAndEscape(text));
         }
@@ -143,25 +132,5 @@ public final class NoneLength extends Length<Void> implements HasJsonNode, Value
     @Override
     public String toString() {
         return TEXT;
-    }
-
-    // HasJsonNode......................................................................................................
-
-    /**
-     * Accepts a json string holding a number and px unit suffix.
-     */
-    public static NoneLength fromJsonNodeNormal(final JsonNode node) {
-        Objects.requireNonNull(node, "node");
-
-        try {
-            return parseNormal(node.stringValueOrFail());
-        } catch (final JsonNodeException cause) {
-            throw new IllegalArgumentException(cause.getMessage(), cause);
-        }
-    }
-
-    @Override
-    public JsonNode toJsonNode() {
-        return JsonNode.string(this.toString());
     }
 }
